Fix crash caused by deleting a ShipDesign.
Code was removing the design from the current ship manager and then
checking (after it was removed) if it was obsolete causing the crash.

Fix by a) moving the removal to after the check and b) adding the check
for isKnown before the check for isObsolete.

Bug found in review #1587.