Add explicit std::invalid_argument parameter to variant of ShipDesign()
The explicit std::invalid_argument distinguishes the throwing variant of
the ShipDesign constructor and might reduce the likelihood of using it
and not catching the exception.

This constructor can be used to detect bad input arguments and produce a
detailed description of the FOCS changes required to correct the
problem.