Modify ShipDesign constructor to either throw or log.
ShipDesign constructor constructs a valid design unless there are no
hulls available.

If invalid arguments are provided it now either throws
std::invalid_argument or produces a detailed log report and adjusts the
input arguments to produce a valid design.

Previously, it would produce the log report even when throwing, which
results in pointless noise in the log when an algorithm is using the
constructor to determine design validity.